Move new BC interfaces into a shared package.
