+++ enterprise/internal/batches/background.go
@@ -8,10 +8,10 @@ import (
 
 	"github.com/sourcegraph/log"
 
-	"github.com/sourcegraph/sourcegraph/cmd/repo-updater/repoupdater"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/batches/store"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/batches/syncer"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
+	"github.com/sourcegraph/sourcegraph/internal/batches"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/encryption"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
@@ -27,7 +27,7 @@ func InitBackgroundJobs(
 	db database.DB,
 	key encryption.Key,
 	cf *httpcli.Factory,
-) repoupdater.ChangesetSyncRegistry {
+) batches.ChangesetSyncRegistry {
 	// We use an internal actor so that we can freely load dependencies from
 	// the database without repository permissions being enforced.
 	// We do check for repository permissions consciously in the Rewirer when
